RPG-1896 API: Stop looping through every item on the view all the time char moves (#521)
* RPG-1896 Added debounce function to avoid unnecessary function calls

* RPG-1896 feat: Add logic to prevent re-adding items to character view when already present

This commit updates the 'warnCharacterAboutItemsInView' method in the 'CharacterMovementWarn.ts' class to prevent re-adding items to the character's view when they are already present. The new logic checks if the item is already in the view before adding it, which avoids unnecessary updates and improves performance.

The behavior can be toggled using an optional 'always' flag in the method's options parameter, which defaults to false. When set to false, items are only added to the view if they are not already present. When set to true, items are added regardless of whether they are already in the view.

---------

Co-authored-by: Joao Paulo Furtado <joaopaulofurtado@live.com>